pyproject.toml
[tool.poetry]
name = "finance-sentiment-tracker"
version = "0.1.0"
description = ""
authors = ["Luis Peixoto <lmpeixoto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nltk = "^3.9.1"
newspaper4k = "^0.9.3.1"
lxml-html-clean = "^0.4.1"
typing-extensions = "^4.12.2"
python-decouple = "^3.8"
datetime = "^5.5"
sqlalchemy = "^2.0.36"
flake8 = "^7.1.1"
black = "^24.10.0"
pre-commit = "^4.0.1"
fastapi = "^0.115.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
